build: enable compilation with Clang/Apple Clang
Two Clang-only build breakers in the linear-algebra/box code (GCC accepts
both, so the GCC-only CI never caught them):

- Vector3D compound-assignment operators declared a parenthesized
  requires(...) clause but defined an unparenthesized one; constraint
  expressions must be token-identical, so Clang rejects the out-of-line
  definitions.
- TriclinicBox::applyPBC built Vec3D{i, j, k} from int loop variables, a
  narrowing conversion in a braced initializer (ill-formed; GCC only warns).

No behavior change on GCC.

+++ src/box/triclinicBox.cpp
@@ -128,7 +128,8 @@ void TriclinicBox::applyPBC(Vec3D &position) const
             for (int j = -1; j <= 1; ++j)
                 for (int k = -1; k <= 1; ++k)
                 {
-                    const auto shift = _boxMatrix * Vec3D{i, j, k};
+                    const auto shift =
+                        _boxMatrix * Vec3D{double(i), double(j), double(k)};
 
                     const auto newPosition = originalPosition + shift;
